[closure segmenter] don't reprocess inert segments.
Segments that are found to only have exclusive gids are inert. They don't interact with any other segments. We can utilize this to reduce the amount of closure operations needed. When merging a set of segments where all segments are inert then there is no need to do closure analysis as we can form the glyph set of the merged segment by just unioning all of the exclusive gids.

This updates the closure segmenter to track which segments are inert and skip closure analysis for those.