Don't use subtraction when computing the 'except' segments.
See: https://github.com/w3c/ift-encoder/blob/main/docs/experimental/closure_glyph_segmentation.md

The 'except' segment is defined as the union of all other segments except for the one being analyzed. As a short cut we were computing it by subtracting the segment of interest from the union of all segments. However, this is not correct when there are overlaps between segments as subtraction may remove the overlapping parts from the except segment. Change to computing the except segment as defined in the design doc.

Adds tests for cases that run into this issue.