Refactor GlyphGroupings to use separate exclusive glyphs grouping.
Prior to this change exclusive glyphs were being kep in the 'and glyph' groupings and down stream users needed to know that 'and glyph' groupings of size 1 are exclusive patches. This change moves exclusive groupings to their own map to make it clear and reduce the need for special casing in various other places.