Rework closure segmenter to support more than one merge group.
Segment merging will be restricted to be within the merge group. As a result merges won't be considered that cross merge groups. This will allow segmenting to be performed for multiple scripts that each have their own frequency data. For now segments that appear in two or more merge groups are separated off and merging is disabled for them. More work is needed to handle these shared segments.