Rework the analysis for moving patches to the initial font to be purely glyph based.
When working with the initial font we now have the ability to directly specify the exact set of glyphs in the initial font. This updates the initial font moving analysis to take advantage of this. Using glyphs allows us to move just individual patches into the initial font without pulling in unintended stuff via closure which can happen when working with codepoints only. Working with glyphs also allows for a simpler and more accurate delta computation.