Fix bug with inert segment processing.
The implementation was making the assumption that if two inert segments are combined that the new segment must also be inert. It turns out this isn't always true. For example:
- Let's say we have a conjunctive conditions which is detected as:

  s10 AND ... additional conditions

  where the true undiscovered condition is:

  s10 AND (s1 OR s2)

- Then if s1 and s2 are inert, upon combining them, the additional condition (s1 OR s2) will become discoverable and the new condition mapping is:

  if s1+2 -> ...
  if s10 AND s1+2 -> ...

  Therefore the new combined segment s1+2 is no longer inert.

To fix this we always check for inertness of the merged segement during merge application. For merge analysis the logic is unchanged since it is still OK to assume the merged patch produced from inert segments will be the combination of the input segment glyphs.