For best case size computation add a configurable reduction percentage.
The merger prunes patch pairs from evaluation by looking at a "best case" cost reduction. During
this calculation the size of the merged patch is estimated using the size of the individual patches
being merged. The following forumula is used:

merged_size =
  largest_individual_patch_size +
  best_case_size_reduction_fraction * sum of remaining patch sizes

Setting best_case_size_reduction_fraction to 0 gives the previous behaviour, larger values cause
more aggressive pruning, at the cost of possibly underestimating the true best case for a pair,
and over pruning pairs that may have actually been viable.

Also adds a flag that when enabled has the merger record a histogram of encountered size reductions.
This was used on a few example fonts to inform the default value of 0.50. In testing the default value gives
a significant speedup, while having neglible impact on the total cost of produced segmentations.