Initial implementation of complex condition detection.
This is used to assign conditions to unmapped (fallback) glyphs that fail to have conditions detected by regular closure analysis. The complex detection tries to find a minimal purely disjunctive condition for each unmapped glyph. This condition is not the true condition for the glyph but is a minimal purely disjunctive superset of the true condition.

Currently this is done as a post-merging step, however I'm ultimately aiming to have this done pre-merging and make merging capable of incrementally applying complex condition detection as needed.

This adds a new field to the segmenter config to control how unmapped glyphs are handled, there are three options:
- Move them to the initial font (existing default behaviour)
- Place them in an always loaded patch.
- Run the newly added complex condition detection and place in more granular patches based on that.

Lastly, no tests have been added yet. Those are coming next.